fix the import logic, fix tests, put all tensorboard in a single folder
